requirements.txt
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
future==0.16.0
idna==2.6
itchat==1.2.32
matplotlib==2.1.2
numpy==1.14.0
Pillow==5.0.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wxpy==0.3.9.8
